requirements.txt
streamlit>=1.33.0
pandas>=2.0.0
openpyxl>=3.1.0
bcrypt>=4.0.0
pyyaml>=6.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
